+++ frontend/src/mod/algorithm/parserTools/test/parser.test.ts
@@ -1,9 +1,11 @@
 import { expect, test } from 'vitest';
 import { Polygon } from '@/mod/data/polygon/polygon';
 import { Room } from '@/mod/data/room/room';
-import * as polygonTools from './polygonTools';
-import * as roomTools from './roomTools';
-import * as graphTools from './graphTools';
+import * as polygonTools from '../polygonTools';
+import * as roomTools from '../roomTools';
+import * as graphTools from '../pathTools';
+
+import type { Line } from '@/mod/data/com/line';
 
 const testPolygon = new Polygon(
     [
@@ -38,7 +40,7 @@ const testRoom2 = new Room(7107, 1, testPolygon2, 'testRoom2ID');
 const testRoom3 = new Room(7109, 1, testPolygon3, 'testRoom3ID');
 
 test('polygon edge extraction', () => {
-    const edges: polygonTools.Line[] = polygonTools.getPolygonEdges(testPolygon);
+    const edges: Line[] = polygonTools.getPolygonEdges(testPolygon);
     expect(edges).toEqual([
         [0, 0],
         [false, 0],
@@ -71,7 +73,6 @@ test('room touch extraction', () => {
 });
 
 test('graph generation test', () => {
-    const graph = graphTools.generateGraph([testRoom, testRoom2, testRoom3]);
-    console.log(graph);
+    graphTools.generatePath([testRoom, testRoom2, testRoom3]);
     expect(true).toBe(true);
 });